src/cloud/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
google-cloud-storage==2.10.0
functions-framework==3.4.0